feat: Edit subject from Search section
Handle 'Edit this subject' from search submenu. Now when user choose
this option could update the subject consulted.

On the other hand, it was necessary to add a mini version of search
submenu, because it's necessary a subject to use the option 'Edit this
subject' or 'Remove this subject'. This version has the following options:

- Search another subject
- Back to menu

Resolve: #5
View also: #1